[fix] Institue,Teacher, Student Controller 유효성 검사 추가 및 수정 , Student RepositoryImpl 방식으로 변경

+++ src/main/java/com/TeachingManager/TeachingManager/controller/API/Institute/ApiInstituteController.java
@@ -10,6 +10,7 @@ import com.TeachingManager.TeachingManager.domain.Institute;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -29,16 +30,22 @@ public class ApiInstituteController {
 
     // 학원 정보 수정
     @PutMapping("/api/institute") ResponseEntity<InstituteResponse> update_Institute(@RequestBody UpdateInstituteRequest request, @AuthenticationPrincipal CustomUser user){
-        Institute institute = instService.update_Institute(user, request);
-        return ResponseEntity.ok()
-                .body(new InstituteResponse(institute));
+        if(user != null && user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_PRESIDENT"))) {
+            Institute institute = instService.update_Institute(user, request);
+            return ResponseEntity.ok()
+                    .body(new InstituteResponse(institute));
 
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     // 회원탈퇴 api
     @PutMapping("/api/delete/institute") ResponseEntity<String> delete_Institute(@AuthenticationPrincipal CustomUser user){
-        return ResponseEntity.ok()
-                .body(instService.delete_Institute(user));
+        if(user != null && user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_PRESIDENT"))) {
+            return ResponseEntity.ok()
+                    .body(instService.delete_Institute(user));
+        }
+        return ResponseEntity.badRequest().build();
     }
 
 }
