+++ src/main/java/com/TeachingManager/TeachingManager/controller/API/Student/ApiStudentController.java
@@ -1,7 +1,6 @@
 package com.TeachingManager.TeachingManager.controller.API.Student;
 
 import com.TeachingManager.TeachingManager.Service.Student.StudentService;
-import com.TeachingManager.TeachingManager.Service.User.TokenService;
 import com.TeachingManager.TeachingManager.domain.CustomUser;
 import com.TeachingManager.TeachingManager.domain.Student;
 import com.TeachingManager.TeachingManager.DTO.Student.AddStudentRequest;
@@ -11,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,55 +23,54 @@ public class ApiStudentController {
     // 학생 추가 api
     @PostMapping("/api/students")
     public ResponseEntity<Student> addStudent(@AuthenticationPrincipal CustomUser user,@RequestBody AddStudentRequest request){
-        // 헤더에서 추출한 이메일 값으로 생성하기
-        Student savedStudent = studentService.create_student(request, user.getPk());
-        if (savedStudent != null){
+        if(user != null && user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_PRESIDENT"))) {
             return ResponseEntity.status(HttpStatus.CREATED)
-                .body(savedStudent);
+                        .body(studentService.create_student(request, user));
         }
-        else{
-            // 존재하지 않는 이메일로 들어왔을 경우.
-            return (ResponseEntity<Student>) ResponseEntity.status(HttpStatus.BAD_REQUEST);
-        }
+        return ResponseEntity.badRequest().build();
     }
 
     // 전체 학생 조회
     @GetMapping("/api/students")
     public ResponseEntity<List<StudentResponse>> findAllStudents(@AuthenticationPrincipal CustomUser user) {
-        List<StudentResponse> students = studentService.findAll(user)
-                .stream()
-                .map(StudentResponse::new)
-                .toList();
-
-        return ResponseEntity.ok()
-                .body(students);
+        if(user != null) {
+            return ResponseEntity.ok()
+                    .body(studentService.findAll(user));
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     // 단일 학생 조회
     @GetMapping("/api/students/{id}")
     public ResponseEntity<StudentResponse> findStudent(@AuthenticationPrincipal CustomUser user, @PathVariable(name = "id") long id){
-        Student student = studentService.findById(user, id);
-        return ResponseEntity.ok()
-                .body(new StudentResponse(student));
+        if(user != null) {
+            Student student = studentService.findById(user, id);
+            return ResponseEntity.ok()
+                    .body(new StudentResponse(student));
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     //  삭제. 삭제시 삭제된 학생 이름 전달.
     @PostMapping("/api/delete/students/{id}")
     public ResponseEntity<String> deleteArticle(@AuthenticationPrincipal CustomUser user,@PathVariable(name="id") long id){
-        String student_name = studentService.delete(user, id);
-
-        return ResponseEntity.ok()
-                .body(student_name);
+        if(user != null && user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_PRESIDENT"))) {
+            String student_name = studentService.delete(user, id);
+            return ResponseEntity.ok()
+                    .body(student_name);
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     // 학생 업데이트
     @PutMapping("/api/students/{id}")
     public ResponseEntity<StudentResponse> updateStudent(@AuthenticationPrincipal CustomUser user, @PathVariable(name = "id") long id, @RequestBody UpdateStudentRequest request) {
-        Student updatedStudent = studentService.update(user, id, request);
-
-        return ResponseEntity.ok()
-                .body(new StudentResponse(updatedStudent));
-
+        if(user != null && user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_PRESIDENT"))) {
+            Student updatedStudent = studentService.update(user, id, request);
+            return ResponseEntity.ok()
+                    .body(new StudentResponse(updatedStudent));
+        }
+        return ResponseEntity.badRequest().build();
     }
 
 }
